requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
get==1.0.3
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-language==1.0.2
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.12.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
post==1.0.2
protobuf==3.6.0
public==1.0.3
pyasn1==0.4.3
pyasn1-modules==0.2.1
PySocks==1.6.8
pytz==2018.4
query-string==1.0.2
request==1.0.2
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
tweepy==3.6.0
urllib3==1.23
Werkzeug==0.14.1
